apps/api/requirements.txt
fastapi==0.115.12
open_clip_torch==2.24.0
pillow==11.2.1
scikit-learn==1.4.2
timm==1.0.15
torch==2.7.1
torchvision==0.22.1
uvicorn==0.34.3
